Fixing merge issues
