--- src/app/features/workplace/view-user-account/view-user-account.component.html
@@ -1,76 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-one-third-from-desktop">
-    <span class="govuk-caption-l">User Account</span>
-    <h1 class="govuk-heading-l">{{ user.fullname }}</h1>
-  </div>
-  <div class="govuk-grid-column-two-thirds-from-desktop">
-    <div class="govuk-button-heading-group">
-      <button *ngIf="canResendActivationLink" class="govuk-button govuk-button--link" (click)="resendActivationLink()">
-        Resend account set up link
-      </button>
-      <button *ngIf="canDeleteUser" class="govuk-button govuk-button--link" (click)="onDeleteUser()">
-        <img src="/assets/images/bin.svg" />
-        <span class="govuk-!-margin-left-1">
-          Delete <span class="govuk-visually-hidden">{{ user.fullname }}'s</span> user account
-        </span>
-      </button>
-    </div>
-  </div>
-</div>
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-full">
-    <app-summary-list [topBorder]="true" [summaryList]="userInfo"></app-summary-list>
-    <app-summary-list [topBorder]="true" [summaryList]="loginInfo"></app-summary-list>
-    <form (ngSubmit)="onSubmit()" [formGroup]="form">
-      <dl
-        class="govuk-summary-list govuk-summary-list--no-border govuk-summary-list--wrap-border govuk-!-margin-bottom-0"
-      >
-        <div class="govuk-summary-list__row">
-          <dt class="govuk-summary-list__key">
-            Permissions
-          </dt>
-          <dd class="govuk-summary-list__value">
-            <div class="govuk-radios">
-              <ng-container *ngFor="let item of roleRadios; let i = index">
-                <div class="govuk-radios__item">
-                  <input
-                    class="govuk-radios__input"
-                    [formControlName]="'role'"
-                    id="role-{{ i }}"
-                    name="role"
-                    value="{{ item.value }}"
-                    type="radio"
-                  />
-                  <label class="govuk-label govuk-radios__label" for="role-{{ i }}">
-                    {{ item.label }}
-                  </label>
-                </div>
-                <div
-                  *ngIf="item.value === 'Edit'"
-                  class="govuk-radios__conditional"
-                  [class.govuk-radios__conditional--hidden]="form.get('role').value !== 'Edit'"
-                  id="role-conditional-{{ i }}"
-                >
-                  <div class="govuk-checkboxes">
-                    <div class="govuk-checkboxes__item">
-                      <input
-                        class="govuk-checkboxes__input"
-                        id="primary"
-                        name="primary"
-                        type="checkbox"
-                        [formControlName]="'primary'"
-                      />
-                      <label class="govuk-label govuk-checkboxes__label" for="primary">
-                        Make this user a Primary User
-                      </label>
-                    </div>
-                  </div>
-                </div>
-              </ng-container>
-            </div>
-          </dd>
-        </div>
-      </dl>
-    </form>
-  </div>
-</div>
